pyproject.toml
[tool.poetry]
name = "platform-examples"
version = "0.1.0"
description = "Simple examples that express platform engineering"
authors = ["Chris Regnier <christopher.regnier@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src/python"}]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
aiosqlite = "^0.20.0"
python-decouple = "^3.8"
uvicorn = "^0.27.1"
greenlet = "^3.0.3"
sqlalchemy = "^2.0.28"
pytest = "^8.0.2"
polyfactory = "^2.15.0"
pytest-asyncio = "^0.23.5.post1"
pytest-coverage = "^0.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = ["--cov=src/python", "--cov-report=term-missing", "--cov-report=html:tests/python/reports/html/", "--cov-report=xml:tests/python/reports/coverage.xml", "--cov-branch"]
testpaths = ["tests/python"]

